+++ kaithem/src/docs/changes.md
@@ -20,6 +20,8 @@ Change Log
 - Minor breaking: Resources all have file extensions, old loaded modules may have odd names but will load
 - Events are now stored as standard python files with data in variables, for easy viewing in external editors
 - New kaithem.web.controllers: Easily create pages directly in python code using cherrypy directly without losing the flexibility of Kaithem.
+- WiFi Manager, on Linux with NetworkManager you can set up connections to access points via
+  the web UI.
 
 
 ### 0.63.1